workers/ingestion/requirements.txt
# RabbitMQ
aio-pika==9.3.1

# Database
asyncpg==0.29.0

# Qdrant
qdrant-client==1.7.3

# Azure
azure-ai-formrecognizer==3.3.2
openai==1.10.0

# Utilities
python-dotenv==1.0.0
aiohttp==3.9.1
